+++ src/command_modules/azure-cli-billing/azure/cli/command_modules/billing/_help.py
@@ -19,3 +19,8 @@ helps['billing period'] = """
     type: group
     short-summary: Get billing periods for a subscription.
 """
+
+helps['billing enrollment-account'] = """
+    type: group
+    short-summary: Get enrollment accounts.
+"""