+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_validators.py
@@ -59,9 +59,25 @@ def validate_nsg_name(cmd, namespace):
         or '{}_NSG_{}'.format(namespace.vm_name, hash_string(vm_id, length=8))
 
 
-def process_vm_secret_namespace(cmd, namespace):
+def validate_keyvault(cmd, namespace):
     namespace.keyvault = _get_resource_id(cmd.cli_ctx, namespace.keyvault, namespace.resource_group_name,
                                           'vaults', 'Microsoft.KeyVault')
+
+
+def process_vm_secret_format(cmd, namespace):
+    from msrestazure.tools import is_valid_resource_id
+
+    keyvault_usage = CLIError('usage error: [--keyvault NAME --resource-group NAME | --keyvault ID]')
+    kv = namespace.keyvault
+    rg = namespace.resource_group_name
+
+    if rg:
+        if not kv or is_valid_resource_id(kv):
+            raise keyvault_usage
+        validate_keyvault(cmd, namespace)
+    else:
+        if kv and not is_valid_resource_id(kv):
+            raise keyvault_usage
 
 
 def _get_resource_group_from_vault_name(cli_ctx, vault_name):
@@ -88,8 +104,16 @@ def _get_resource_id(cli_ctx, val, resource_group, resource_type, resource_names
     if is_valid_resource_id(val):
         return val
 
-    return resource_id(name=val, resource_group=resource_group, namespace=resource_namespace, type=resource_type,
-                       subscription=get_subscription_id(cli_ctx))
+    kwargs = {
+        'name': val,
+        'resource_group': resource_group,
+        'namespace': resource_namespace,
+        'type': resource_type,
+        'subscription': get_subscription_id(cli_ctx)
+    }
+    missing_kwargs = {k: v for k, v in kwargs.items() if not v}
+
+    return resource_id(**kwargs) if not missing_kwargs else None
 
 
 def _get_nic_id(cli_ctx, val, resource_group):
@@ -266,6 +290,19 @@ def _validate_managed_disk_sku(sku):
     allowed_skus = ['Premium_LRS', 'Standard_LRS']
     if sku and sku.lower() not in [x.lower() for x in allowed_skus]:
         raise CLIError("invalid storage SKU '{}': allowed values: '{}'".format(sku, allowed_skus))
+
+
+def _validate_location(cmd, namespace, zone_info, size_info):
+    from ._vm_utils import list_sku_info
+    if not namespace.location:
+        get_default_location_from_resource_group(cmd, namespace)
+        if zone_info:
+            sku_infos = list_sku_info(cmd.cli_ctx, namespace.location)
+            temp = next((x for x in sku_infos if x.name.lower() == size_info.lower()), None)
+            if not temp or not [x for x in (temp.location_info or []) if x.zones]:
+                raise CLIError("{}'s location can't be used to create the VM/VMSS because availablity zone is not yet "
+                               "supported. Please use '--location' to specify a capable one. 'az vm list-skus' can be "
+                               "used to find such locations".format(namespace.resource_group_name))
 
 
 # pylint: disable=too-many-branches, too-many-statements
@@ -862,7 +899,7 @@ def _resolve_role_id(cli_ctx, role, scope):
 
 def process_vm_create_namespace(cmd, namespace):
     validate_tags(namespace)
-    get_default_location_from_resource_group(cmd, namespace)
+    _validate_location(cmd, namespace, namespace.zone, namespace.size)
     validate_asg_names_or_ids(cmd, namespace)
     _validate_vm_create_storage_profile(cmd, namespace)
     if namespace.storage_profile in [StorageProfile.SACustomImage,
@@ -1033,7 +1070,7 @@ def get_network_lb(cli_ctx, resource_group_name, lb_name):
 
 def process_vmss_create_namespace(cmd, namespace):
     validate_tags(namespace)
-    get_default_location_from_resource_group(cmd, namespace)
+    _validate_location(cmd, namespace, namespace.zones, namespace.vm_sku)
     _validate_vm_create_storage_profile(cmd, namespace, for_scale_set=True)
     _validate_vm_vmss_create_vnet(cmd, namespace, for_scale_set=True)
     _validate_vmss_create_load_balancer_or_app_gateway(cmd, namespace)