+++ src/command_modules/azure-cli-profile/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 2.0.22
 ++++++
+* account list: handle accounts which come from CLI 1.0/ASM mode
 * (Breaking change): remove `--msi` & `--msi-port` which were tagged `deprecating` 2 releases ago
 
 2.0.21